+++ myems-admin/app/controllers/fdd/rule/rule.controller.js
@@ -10,6 +10,8 @@ app.controller('RuleController', function(
     toaster,
     SweetAlert) {
     $scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
+	$scope.exportdata = '';
+	$scope.importdata = '';
 
 	$scope.initExpression = [{
 		"sample_object_id": 1,
@@ -175,6 +177,97 @@ app.controller('RuleController', function(
             }
         });
     };
+
+	$scope.exportRule = function(rule) {
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		RuleService.exportRule(rule, headers, function(response) {
+			if (angular.isDefined(response.status) && response.status === 200) {
+				$scope.exportdata = JSON.stringify(response.data);
+				var modalInstance = $uibModal.open({
+					windowClass: "animated fadeIn",
+					templateUrl: 'views/common/export.html',
+					controller: 'ModalExportCtrl',
+					resolve: {
+						params: function() {
+							return {
+								exportdata: angular.copy($scope.exportdata)
+							};
+						}
+					}
+				});
+				modalInstance.result.then(function() {
+					//do nothing;
+				}, function() {
+					//do nothing;
+				});
+				$rootScope.modalInstance = modalInstance;
+			} else {
+				$scope.exportdata = null;
+			}
+		});
+	};
+
+	$scope.cloneRule = function(rule){
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		RuleService.cloneRule(rule, headers, function(response) {
+			if (angular.isDefined(response.status) && response.status === 201) {
+				toaster.pop({
+					type: "success",
+					title: $translate.instant("TOASTER.SUCCESS_TITLE"),
+					body: $translate.instant("TOASTER.SUCCESS_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+					showCloseButton: true,
+				});
+				$scope.getAllRules();
+				$scope.$emit('handleEmitRuleChanged');
+			}else {
+				toaster.pop({
+					type: "error",
+					title: $translate.instant("TOASTER.ERROR_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+					body: $translate.instant(response.data.description),
+					showCloseButton: true,
+				});
+			}
+		});
+	};
+
+	$scope.importRule = function() {
+		var modalInstance = $uibModal.open({
+			templateUrl: 'views/common/import.html',
+			controller: 'ModalImportCtrl',
+			windowClass: "animated fadeIn",
+			resolve: {
+				params: function() {
+					return {
+					};
+				}
+			}
+		});
+		modalInstance.result.then(function(importdata) {
+			let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+			RuleService.importRule(importdata, headers, function(response) {
+				if (angular.isDefined(response.status) && response.status === 201) {
+					toaster.pop({
+						type: "success",
+						title: $translate.instant("TOASTER.SUCCESS_TITLE"),
+						body: $translate.instant("TOASTER.SUCCESS_ADD_BODY", {template: $translate.instant("COMMON.DATA_SOURCE")}),
+						showCloseButton: true,
+					});
+					$scope.getAllRules();
+					$scope.$emit('handleEmitRuleChanged');
+				} else {
+					toaster.pop({
+						type: "error",
+						title: $translate.instant("TOASTER.ERROR_ADD_BODY", { template: $translate.instant("COMMON.DATA_SOURCE") }),
+						body: $translate.instant(response.data.description),
+						showCloseButton: true,
+					});
+				}
+			});
+		}, function() {
+
+		});
+		$rootScope.modalInstance = modalInstance;
+	};
 
 	$scope.getAllRules();